bot/requirements.txt
ccxt==4.2.14
pandas==2.2.0
python-dotenv==1.0.1
schedule==1.2.1
requests==2.31.0
yfinance==0.2.54
